requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
ffmpeg-python
werkzeug
python-dotenv
celery
redis
